Feature :: Add multilingual in the system.
There will be one default language and multiple supported language (inlcuding default).
Default language can be changed from the list of supported language. And language can be added and removed from the system.
If any language is removed then all the data of that language will be removed from the system and if any new language is added then default langugae data will be copied with new lang_code.
Prioiry of language:-
[1] lang_code from URL
[2] lang_code in post data
[3] Joomla's current language (is supported)
[4] Paycart's default language
Whenever a new record is created, its data will be copied of the other languages too.

ref #270,#348